Merge commit 8c38460eba008fd72aff6a4f0febf25d7fc2eba5 from master.
Fix scanning of 0 columns from external table. Fixes #63.

The previous solution involved registering a DataFrame with 0 columns as a temporary table and then running a select * from it. This works in spark 1.5.1, but not with 1.6.x.

I raised https://issues.apache.org/jira/browse/SPARK-16329, but we can just avoid doing that altogether by calling DataFrameWriter.insertInto() instead. That's what this patch is about.