External Scan Predicate Pushdown
The corresponding x100 changes extended JobPart with per-column "value ranges" that describe the selection predicate which can be pushed down.

This change makes the provider capable of interpreting these textual value ranges and turning them into a boolean expression (in Spark terms), which is then applied as a filter to the incoming data DF inside the insert() method of the Insertable VectorRelation.